+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `clear_standalone!` to components
 - Allow passing an array to `prevent` to prevent multiple actions at once
 - Add features `Compony.content_before_root_comp` and `Compony.content_after_root_comp`
+- Add new field kind percentage
 
 # 0.0.8